Model loading (.obj)

+++ Avalanche/Source/Engine/Vulkan/BufferBase.h
@@ -6,7 +6,7 @@
 class Buffer
 {
 public:
-    Buffer(vk::BufferUsageFlags usage, const void* bufferData, uint32_t size);
+    Buffer(vk::BufferUsageFlags usage, const void* bufferData, size_t size);
     ~Buffer();
 
     const vk::Buffer& GetBuffer() const { return m_Buffer; }
